🚚 Rename Configuration as FilterConfiguration
