+++ cumplo_common/database/firestore.py
@@ -7,7 +7,7 @@ from google.cloud.firestore_v1.base_query import FieldFilter
 from google.cloud.firestore_v1.document import DocumentReference
 
 from cumplo_common.models.channel import CHANNEL_CONFIGURATION_BY_TYPE, ChannelConfiguration, ChannelType
-from cumplo_common.models.configuration import Configuration
+from cumplo_common.models.filter import FilterConfiguration
 from cumplo_common.models.notification import Notification
 from cumplo_common.models.user import User
 from cumplo_common.utils.constants import (
@@ -64,14 +64,14 @@ class FirestoreClient:
         if not (user_data := user.to_dict()):
             raise ValueError(f"User with API key {secure_key(api_key)} data is empty")
 
-        configurations = self._get_user_configurations(user.id)
-        notifications = self._get_user_notifications(user.id)
+        filters = self._get_user_filters(user.id)
         channels = self._get_user_channels(user.id)
+        notifications = self._get_user_notifications(user.id)
         return User(
             id=user.id,
-            configurations=configurations,
-            notifications=notifications,
+            filters=filters,
             channels=channels,
+            notifications=notifications,
             **user_data,
         )
 
@@ -87,7 +87,7 @@ class FirestoreClient:
         notification = self._get_notification_document(id_user, id_funding_request)
         notification.set({"date": arrow.utcnow().datetime})
 
-    def put_configuration(self, id_user: str, configuration: Configuration) -> None:
+    def put_filter(self, id_user: str, configuration: FilterConfiguration) -> None:
         """
         Creates or updates a configuration of a given user
 
@@ -96,8 +96,8 @@ class FirestoreClient:
             configuration (Configuration): The Configuration object containing the new configuration data to be updated
         """
         logger.info(f"Updating configuration {configuration.id} of user {id_user} at Firestore")
-        configuration_reference = self._get_configuration_document(id_user, configuration.id)
-        configuration_reference.set(configuration.serialize(for_firestore=True))
+        configuration_reference = self._get_filter_document(id_user, configuration.id)
+        configuration_reference.set(configuration.json())
 
     def put_channel(self, id_user: str, channel: ChannelConfiguration) -> None:
         """
@@ -123,16 +123,16 @@ class FirestoreClient:
         notification = self._get_notification_document(id_user, id_funding_request)
         notification.delete()
 
-    def delete_configuration(self, id_user: str, id_configuration: int) -> None:
+    def delete_filter(self, id_user: str, id_filter: int) -> None:
         """
         Deletes a configuration for a given user and configuration ID
 
         Args:
             id_user (str): The user ID which owns the configuration
-            id_configuration (int): The configuration ID to be deleted
+            id_filter (int): The filter ID to be deleted
         """
-        logger.info(f"Deleting configuration {id_configuration} from Firestore")
-        configuration = self._get_configuration_document(id_user, id_configuration)
+        logger.info(f"Deleting filter {id_filter} from Firestore")
+        configuration = self._get_filter_document(id_user, id_filter)
         configuration.delete()
 
     def delete_channel(self, id_user: str, channel_type: ChannelType) -> None:
@@ -156,10 +156,10 @@ class FirestoreClient:
         user_document = self._get_user_document(id_user)
         return user_document.collection(NOTIFICATIONS_COLLECTION).document(str(id_funding_request))
 
-    def _get_configuration_document(self, id_user: str, id_configuration: int) -> DocumentReference:
+    def _get_filter_document(self, id_user: str, id_filter: int) -> DocumentReference:
         """Gets a configuration document reference"""
         user_document = self._get_user_document(id_user)
-        return user_document.collection(CONFIGURATIONS_COLLECTION).document(str(id_configuration))
+        return user_document.collection(CONFIGURATIONS_COLLECTION).document(str(id_filter))
 
     def _get_channel_document(self, id_user: str, channel_type: ChannelType) -> DocumentReference:
         """Gets a channel document reference"""
@@ -175,14 +175,14 @@ class FirestoreClient:
         notifications = user_document.collection(NOTIFICATIONS_COLLECTION).stream()
         return {int(n.id): Notification(id=int(n.id), **n.to_dict()) for n in notifications}
 
-    def _get_user_configurations(self, id_user: str) -> dict[int, Configuration]:
+    def _get_user_filters(self, id_user: str) -> dict[int, FilterConfiguration]:
         """
         Gets the user configurations data
         """
         logger.info(f"Getting user {id_user} configurations from Firestore")
         user_document = self._get_user_document(id_user)
         configurations = user_document.collection(CONFIGURATIONS_COLLECTION).stream()
-        return {int(c.id): Configuration(id=int(c.id), **c.to_dict()) for c in configurations}
+        return {int(c.id): FilterConfiguration(id=int(c.id), **c.to_dict()) for c in configurations}
 
     def _get_user_channels(self, id_user: str) -> dict[ChannelType, ChannelConfiguration]:
         """
